added an overwrite check

+++ action_generate_kibom_templates.py
@@ -2,9 +2,11 @@
 import os
 import shutil
 
-def go_through_directories():
+def go_through_directories(**kwargs):
 
     template_file = 'templates/kibot_template.yaml'
+    report_file = 'working_report.txt'
+    overwrite = kwargs.get('overwrite', False)
     count = 0
     # go through all directories in projects
     for root, dirs, files in os.walk("projects"):
@@ -25,32 +27,34 @@ def go_through_directories():
                     #check that the file with kicad.sch also exists
                     filename_sch = filename.replace(".kicad_pcb",".kicad_sch")
                     if os.path.isfile(filename_sch):
-                        print(f"##############################################")
-                        print(f"###### kibot for {filename}")
-                        ###### copy template across
-                        
-                        filename_directory = os.path.dirname(filename)
-                        output_template = f'{filename_directory}/working_kibot_template.yaml'
-                        #copy and overwrite the template file if it exists
-                        shutil.copyfile(template_file, output_template)
-                        
+                        #if overwrite or report_file doesn't exist
+                        if overwrite or not os.path.isfile(report_file):
+                            print(f"##############################################")
+                            print(f"###### kibot for {filename}")
+                            ###### copy template across
+                            
+                            filename_directory = os.path.dirname(filename)
+                            output_template = f'{filename_directory}/working_kibot_template.yaml'
+                            #copy and overwrite the template file if it exists
+                            shutil.copyfile(template_file, output_template)
+                            
 
-                        ###### generate
+                            ###### generate
 
-                        filename_directory = os.path.dirname(filename)
-                        template = "working_kibot_template.yaml"
-                        #launch using subprocess kibot -c {template} with the base directory being filename_directory
-                        #print(f"working on {filename}")
-                        import subprocess
-                        subprocess.run(["kibot", "-e", "working.kicad_sch", "-b", "working.kicad_pcb", "-c", template], cwd=filename_directory)
+                            filename_directory = os.path.dirname(filename)
+                            template = "working_kibot_template.yaml"
+                            #launch using subprocess kibot -c {template} with the base directory being filename_directory
+                            #print(f"working on {filename}")
+                            import subprocess
+                            subprocess.run(["kibot", "-e", "working.kicad_sch", "-b", "working.kicad_pcb", "-c", template], cwd=filename_directory)
 
-                        
-                        
-                        
-                        count += 1
-                        #print a dot every 100 times through#
-                        if count % 100 == 0:
-                            print(".", end="", flush=True)
+                            
+                            
+                            
+                            count += 1
+                            #print a dot every 100 times through#
+                            if count % 100 == 0:
+                                print(".", end="", flush=True)
         
     print()
     print(f'yaml file for {count} projects created')
